Changed NativeModule loading to be a two stage, modular process. First step is attribute initialization with a BiConsumer and second is general initialization with a Callable (as it was done by default).

+++ app/src/main/java/jg/sh/SeaHorseInterpreter.java
@@ -26,7 +26,7 @@ import net.percederberg.grammatica.parser.ParserCreationException;
 
 public class SeaHorseInterpreter {
   
-  public static final int VERSION = 1;
+  public static final long VERSION = 1;
   
   /**
    * Folder - to be located in the current working directory - that
